requirements.txt
Django==4.2.7
psycopg2-binary==2.9.7
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.5.0